Added better room generation using a brute force algorithm.
The new aglorithm generates pre connected rectangles, and does so by:
* Generating a random rectangle and adding it to a list
* For a certain number of tries, choose a rectangle from the possible rectangle list
    * Move a random distance away and generate a new rectangle
    * Check if the generate rectangle is out of bounds or within a certain distance of any
      other rectangle
    * If not, add it to the list, make a connection between the two rectangles
    * If the original rectangle has more than N connections, remove it from the list of starting rectangles

Originally attempted this with a quad tree, but the results were both less accurate and slower.  The above algorithm
looks N^2 at first (well, ((N / 2) * N), but actually ends up being much closer to n log n.  A quad tree ended up being
K(N), where K is a constant dependent on the map size.  In the case of the map I was using, K was ~400 iterations and was
very slow.  While it meant linearly scaling, for values lower than something insanely high, it was significantly slower,
and the technique used was also significantly less accurate.

There is almost certainly a better data structure out there, but this is good enough for now.